RefactorSeries: Added box type and new series definitions

+++ series/definition_test.go
@@ -23,7 +23,7 @@ func (suite *definitionTestSuite) TestSeries_Len() {
 }
 
 func (suite *definitionTestSuite) TestSeries_GetColumn() {
-	column := base.Column{Name: "col", Dtype: base.Float}
+	column := base.Column{Name: "col", Dtype: base.Float64}
 	// assert that Column returned correctly
 	assert.Equal(suite.T(), column, *suite.S.GetColumn())
 }
